Tidy upp the logged user card and removes unnecesary validation
