+++ src/main.css
@@ -252,7 +252,9 @@ h2 {
   color: #2a363b;
   background: none;
   background-color: #fff;
-  box-shadow: 0 3px 6px rgba(0, 0, 0, 0.16), 0 3px 6px rgba(0, 0, 0, 0.23);
+  box-shadow:
+    0 3px 6px rgba(0, 0, 0, 0.16),
+    0 3px 6px rgba(0, 0, 0, 0.23);
   border-radius: 10px;
 }
 .menuline {
@@ -482,7 +484,9 @@ menu button.active {
   color: #e84a5f;
   font-size: 14px;
   margin: 0 35px 20px;
-  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.12), 0 1px 2px rgba(0, 0, 0, 0.24);
+  box-shadow:
+    0 1px 3px rgba(0, 0, 0, 0.12),
+    0 1px 2px rgba(0, 0, 0, 0.24);
 }
 .download:hover {
   background-color: #2a363b;
@@ -688,11 +692,15 @@ button.rel8 {
   min-height: 220px;
   background-color: #fff;
   padding: 20px;
-  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.12), 0 1px 2px rgba(0, 0, 0, 0.24);
+  box-shadow:
+    0 1px 3px rgba(0, 0, 0, 0.12),
+    0 1px 2px rgba(0, 0, 0, 0.24);
   transition: all 0.3s cubic-bezier(0.25, 0.8, 0.25, 1);
 }
 #filters .item:hover {
-  box-shadow: 0 14px 28px rgba(0, 0, 0, 0.25), 0 10px 10px rgba(0, 0, 0, 0.22);
+  box-shadow:
+    0 14px 28px rgba(0, 0, 0, 0.25),
+    0 10px 10px rgba(0, 0, 0, 0.22);
   background-color: #fecea8;
 }
 #filters p {
@@ -1454,20 +1462,26 @@ div[class*="eaRHYv"] {
   line-height: 1.5;
   border-radius: 5px;
   min-width: 60px;
-  box-shadow: 0px 1px 3px 0px rgba(0, 0, 0, 0.2),
-    0px 1px 1px 0px rgba(0, 0, 0, 0.14), 0px 2px 1px -1px rgba(0, 0, 0, 0.12);
+  box-shadow:
+    0px 1px 3px 0px rgba(0, 0, 0, 0.2),
+    0px 1px 1px 0px rgba(0, 0, 0, 0.14),
+    0px 2px 1px -1px rgba(0, 0, 0, 0.12);
   font-size: 12px;
   white-space: nowrap;
 }
 .graphcard {
   margin: 10px 0;
   border-radius: 5px;
-  box-shadow: 0 3px 6px rgba(0, 0, 0, 0.16), 0 3px 6px rgba(0, 0, 0, 0.23);
+  box-shadow:
+    0 3px 6px rgba(0, 0, 0, 0.16),
+    0 3px 6px rgba(0, 0, 0, 0.23);
   background-color: #fff;
   position: relative;
 }
 .graphcard:hover {
-  box-shadow: 0 10px 20px rgba(0, 0, 0, 0.19), 0 6px 6px rgba(0, 0, 0, 0.23);
+  box-shadow:
+    0 10px 20px rgba(0, 0, 0, 0.19),
+    0 6px 6px rgba(0, 0, 0, 0.23);
 }
 .graphcard h2 {
   border-bottom: 1px solid #e84a5f;
@@ -1581,8 +1595,10 @@ div[class*="eaRHYv"] {
   cursor: pointer;
 }
 .graphcardAcc {
-  box-shadow: 0px 1px 3px 0px rgba(0, 0, 0, 0.2),
-    0px 1px 1px 0px rgba(0, 0, 0, 0.14), 0px 2px 1px -1px rgba(0, 0, 0, 0.12);
+  box-shadow:
+    0px 1px 3px 0px rgba(0, 0, 0, 0.2),
+    0px 1px 1px 0px rgba(0, 0, 0, 0.14),
+    0px 2px 1px -1px rgba(0, 0, 0, 0.12);
   border-bottom-right-radius: 4px;
   border-bottom-left-radius: 4px;
   margin-bottom: 5px;
@@ -1768,7 +1784,9 @@ p.legcomp {
 }
 .total {
   padding: 10px 15px;
-  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.16), 0 3px 6px rgba(0, 0, 0, 0.23);
+  box-shadow:
+    0 1px 3px rgba(0, 0, 0, 0.16),
+    0 3px 6px rgba(0, 0, 0, 0.23);
   background-color: #fff;
   border-radius: 5px;
   margin-right: 5px;
@@ -1784,7 +1802,9 @@ p.legcomp {
   text-align: left;
 }
 .total:hover {
-  box-shadow: 0 10px 20px rgba(0, 0, 0, 0.19), 0 6px 6px rgba(0, 0, 0, 0.23);
+  box-shadow:
+    0 10px 20px rgba(0, 0, 0, 0.19),
+    0 6px 6px rgba(0, 0, 0, 0.23);
 }
 .total h4 {
   font-size: 30px;
@@ -2194,7 +2214,9 @@ tspan,
   margin: auto;
   width: 300px;
   background-color: #ffffff;
-  box-shadow: 0 3px 3px rgba(0, 0, 0, 0.12), 0 3px 3px rgba(0, 0, 0, 0.14);
+  box-shadow:
+    0 3px 3px rgba(0, 0, 0, 0.12),
+    0 3px 3px rgba(0, 0, 0, 0.14);
   border-radius: 5px 0 0 5px;
   padding: 190px 20px 30px;
 }
@@ -2203,13 +2225,17 @@ tspan,
   /*background-image: url(images/bgmountain.png);*/
   background-repeat: no-repeat;
 }
+
 .uim_modal.userModal {
-  background-image: url(images/LogoGEB.png);
+  background-image: var(--profile-img);
   background-position: 50% 20%;
   background-repeat: no-repeat;
 }
+
 .uim_modal:hover {
-  box-shadow: 0 10px 20px rgba(0, 0, 0, 0.19), 0 6px 6px rgba(0, 0, 0, 0.23);
+  box-shadow:
+    0 10px 20px rgba(0, 0, 0, 0.19),
+    0 6px 6px rgba(0, 0, 0, 0.23);
 }
 .login input {
   padding: 0 10px;
@@ -2283,7 +2309,9 @@ tspan,
   position: fixed;
   width: 350px;
   background-color: #ffffff;
-  box-shadow: 0 3px 3px rgba(0, 0, 0, 0.12), 0 3px 3px rgba(0, 0, 0, 0.14);
+  box-shadow:
+    0 3px 3px rgba(0, 0, 0, 0.12),
+    0 3px 3px rgba(0, 0, 0, 0.14);
   bottom: 200px;
   right: 0;
   outline: none;
@@ -2291,7 +2319,9 @@ tspan,
   padding-bottom: 15px;
 }
 .newProjectModal:hover {
-  box-shadow: 0 3px 6px rgba(0, 0, 0, 0.16), 0 3px 6px rgba(0, 0, 0, 0.23);
+  box-shadow:
+    0 3px 6px rgba(0, 0, 0, 0.16),
+    0 3px 6px rgba(0, 0, 0, 0.23);
 }
 .addprjbtn {
   margin-top: 10px;
@@ -2339,10 +2369,14 @@ tspan,
   width: auto;
   background-color: #ffffff;
   border-radius: 5px;
-  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.16), 0 3px 6px rgba(0, 0, 0, 0.23);
+  box-shadow:
+    0 1px 3px rgba(0, 0, 0, 0.16),
+    0 3px 6px rgba(0, 0, 0, 0.23);
 }
 .newBiome:hover {
-  box-shadow: 0 10px 20px rgba(0, 0, 0, 0.19), 0 6px 6px rgba(0, 0, 0, 0.23);
+  box-shadow:
+    0 10px 20px rgba(0, 0, 0, 0.19),
+    0 6px 6px rgba(0, 0, 0, 0.23);
 }
 .newBiome .addprjbtn {
   float: none;
@@ -2490,12 +2524,16 @@ h3.warningNote {
   min-height: 220px;
   background-color: #fff;
   padding: 20px;
-  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.12), 0 1px 2px rgba(0, 0, 0, 0.24);
+  box-shadow:
+    0 1px 3px rgba(0, 0, 0, 0.12),
+    0 1px 2px rgba(0, 0, 0, 0.24);
   transition: all 0.3s cubic-bezier(0.25, 0.8, 0.25, 1);
   border-radius: 15px;
 }
 .portcard:hover {
-  box-shadow: 0 7px 14px rgba(0, 0, 0, 0.25), 0 5px 7px rgba(0, 0, 0, 0.22);
+  box-shadow:
+    0 7px 14px rgba(0, 0, 0, 0.25),
+    0 5px 7px rgba(0, 0, 0, 0.22);
 }
 .portCard .button-container {
   display: flex;
