--- src/utils/validations.ts
@@ -1,9 +1,0 @@
-// TODO: Este utilitario puede (y debe) ser eliminado
-/**
- * Validate if a variable is undefined or null
- * @param {any} variable variable to be validated
- *
- * @returns True if the variable is null or undefined, false otherwise
- */
-export const isUndefinedOrNull = (variable: any) =>
-  typeof variable === "undefined" || variable === null;
